requirements.txt
asgiref==3.3.4
Django==3.2.4
django-cors-headers==3.7.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
Pillow==8.2.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
